nextauthって分けて書くのがAppRouterのお作法らしい
