+++ src/app/api/auth/[...nextauth]/authOptions.ts
@@ -1,0 +1,17 @@
+import GoogleProvider from 'next-auth/providers/google';
+
+export const authOptions = {
+  secret: process.env.NEXTAUTH_SECRET,
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID!,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
+    }),
+  ],
+  callbacks: {
+    async signIn({ user }) {
+      const domain = user.email?.split('@')[1];
+      return domain === process.env.AUTHORIZED_DOMAIN;
+    },
+  },
+};
